Extracted Add To Cart form from Details into it's own component
