+++ src/components/AddToCartForm/index.jsx
@@ -1,0 +1,1 @@
+export { default } from "./AddToCartForm";
